# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payfluence API",
    "version": "v1.0.0",
    "description": "\nPayfluencer is a creator monetization platform that connects brands with content creators through performance-based bounties. This API enables both creators and companies to interact with the platform programmatically and manage workflows including bounty listing, content submission, review, and communication.\n\nCore Capabilities:\n- Creator authentication and profile management\n- Company bounty creation with detailed targeting criteria\n- Submission of creator content for review and reward tracking\n- Real-time chat between creators, clients, and platform admins\n- Status tracking for content submissions (Draft, Submitted, Accepted, Paid)\n\nAuthentication:\nAll endpoints require authentication using Supabase JWT tokens. Only Gmail-based authentication is currently supported.\n\nContent Guidelines:\nBounty listings may include content restrictions such as language, age rating, platform, and visibility requirements.\n\nFor support or questions, contact the Payfluencer team at support@payfluencer.xyz.\n",
    "license": {
      "name": "Copyright 2025 Payfluence"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8001"
    }
  ],
  "paths": {
...
    "/api/v1/auth/google": {
      "post": {
        "summary": "Authenticate user with Google ID token",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idToken"
                ],
                "properties": {
                  "idToken": {
                    "type": "string",
                    "description": "Google ID token from client"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "User logged in successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/GoogleAuthResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/auth/google/callback": {
      "get": {
        "summary": "Handle Google OAuth callback",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Google"
          }
        ],
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "User logged in successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/GoogleAuthResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoogleAuthResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User's email"
          },
          "name": {
            "type": "string",
            "description": "User's name"
          },
          "role": {
            "type": "string",
            "description": "User's role"
          },
          "phoneNumber": {
            "type": "string",
            "description": "User's phone number"
          },
          "createdAt": {
            "type": "string",
            "description": "Account creation date"
          },
          "isActive": {
            "type": "boolean",
            "description": "Account status"
          },
          "profilePicture": {
            "type": "string",
            "description": "URL to user's profile picture"
          },
          "isNewUser": {
            "type": "boolean",
            "description": "Whether this is a new user"
          }
        }
      },
...
    }
  },
...
}